mysite/requirements.txt
django
djangorestframework
environs 
mysqlclient
whitenoise
dj-database-url
psycopg[binary]
django-environ
python-dotenv